Small fixes.

+++ src/drivers/win/debugger.cpp
@@ -1770,9 +1770,9 @@ BOOL CALLBACK DebuggerCallB(HWND hwndDlg, UINT uMsg, WPARAM wParam, LPARAM lPara
 		{
 			CheckDlgButton(hwndDlg, IDC_DEBUGGER_BREAK_ON_CYCLES, break_on_cycles ? BST_CHECKED : BST_UNCHECKED);
 			CheckDlgButton(hwndDlg, IDC_DEBUGGER_BREAK_ON_INSTRUCTIONS, break_on_instructions ? BST_CHECKED : BST_UNCHECKED);
-			sprintf(str, "%d", break_cycles_limit);
+			sprintf(str, "%u", (unsigned)break_cycles_limit);
 			SetDlgItemText(hwndDlg, IDC_DEBUGGER_CYCLES_EXCEED, str);
-			sprintf(str, "%d", break_instructions_limit);
+			sprintf(str, "%u", (unsigned)break_instructions_limit);
 			SetDlgItemText(hwndDlg, IDC_DEBUGGER_INSTRUCTIONS_EXCEED, str);
 
 			CheckDlgButton(hwndDlg, IDC_DEBUGGER_BREAK_ON_BAD_OP, FCEUI_Debugger().badopbreak ? BST_CHECKED : BST_UNCHECKED);
